+++ docker-compose.yml
@@ -11,8 +11,6 @@ services:
             ASPNETCORE_HTTP_PORTS: 8080
             ASPNETCORE_HTTPS_PORTS: 8081
             ConnectionStrings__Database: Host=host.docker.internal;Port=5432;Database=halcyon-api;Username=postgres;Password=password
-            ConnectionStrings__RabbitMq: amqp://guest:guest@host.docker.internal:5672
-            ConnectionStrings__Redis: host.docker.internal
             Email__SmtpServer: host.docker.internal
             Serilog__MinimumLevel__Override__Microsoft.AspNetCore.DataProtection: Error
             Serilog__WriteTo__Seq__Args__serverUrl: http://host.docker.internal:5341
@@ -27,10 +25,6 @@ services:
         depends_on:
             postgres:
                 condition: service_healthy
-            rabbitmq:
-                condition: service_healthy
-            redis:
-                condition: service_healthy
 
     postgres:
         image: postgres:17.2
@@ -44,34 +38,6 @@ services:
         restart: always
         healthcheck:
             test: [ 'CMD', 'pg_isready' ]
-            interval: 10s
-            timeout: 5s
-            retries: 3
-
-    rabbitmq:
-        image: rabbitmq:4.0-management
-        hostname: local
-        ports:
-            - 5672:5672
-            - 15672:15672
-        volumes:
-            - rabbitmq:/var/lib/rabbitmq
-        restart: always
-        healthcheck:
-            test: [ 'CMD', 'rabbitmqctl', 'status' ]
-            interval: 10s
-            timeout: 5s
-            retries: 3
-
-    redis:
-        image: redis:7.4
-        ports:
-            - 6379:6379
-        volumes:
-            - redis:/data
-        restart: always
-        healthcheck:
-            test: [ 'CMD', 'redis-cli', 'ping' ]
             interval: 10s
             timeout: 5s
             retries: 3
@@ -101,10 +67,6 @@ services:
 volumes:
     postgres:
         name: halcyon-postgres
-    rabbitmq:
-        name: halcyon-rabbitmq
-    redis:
-        name: halcyon-redis
     maildev:
         name: halcyon-maildev
     seq:
