--- src/Halcyon.Api/Common/Cache/RedisExtensions.cs
@@ -1,18 +1,0 @@
-﻿namespace Halcyon.Api.Common.Cache;
-
-public static class RedisExtensions
-{
-    public static IHostApplicationBuilder AddRedisDistributedCache(
-        this IHostApplicationBuilder builder,
-        string connectionName
-    )
-    {
-        builder.Services.AddStackExchangeRedisCache(options =>
-        {
-            options.Configuration = builder.Configuration.GetConnectionString(connectionName);
-            options.InstanceName = builder.Environment.ApplicationName;
-        });
-
-        return builder;
-    }
-}
