--- src/Halcyon.Api/Common/Database/EntityChanged/EntityChangedEvent.cs
@@ -1,5 +1,0 @@
-﻿using Microsoft.EntityFrameworkCore;
-
-namespace Halcyon.Api.Common.Database.EntityChanged;
-
-public record EntityChangedEvent(string Entity, EntityState State, Guid? Id) { }
