--- src/Halcyon.Api/Common/Database/EntityChanged/IPublishChanges.cs
@@ -1,6 +1,0 @@
-﻿namespace Halcyon.Api.Common.Database.EntityChanged;
-
-public interface IPublishChanges
-{
-    public Guid Id { get; }
-}
