+++ src/Halcyon.Api/Common/Database/IDbSeeder.cs
@@ -1,6 +1,6 @@
 ﻿using Microsoft.EntityFrameworkCore;
 
-namespace Halcyon.Api.Common.Database.Migration;
+namespace Halcyon.Api.Common.Database;
 
 public interface IDbSeeder<in TDbContext>
     where TDbContext : DbContext
