--- src/Halcyon.Api/Common/Messaging/IMessageConsumer.cs
@@ -1,6 +1,0 @@
-﻿namespace Halcyon.Api.Common.Messaging;
-
-public interface IMessageConsumer<TMessage>
-{
-    Task Consume(TMessage message, CancellationToken cancellationToken);
-}
