--- src/Halcyon.Api/Common/Messaging/IMessagePublisher.cs
@@ -1,6 +1,0 @@
-﻿namespace Halcyon.Api.Common.Messaging;
-
-public interface IMessagePublisher
-{
-    Task Publish<T>(IEnumerable<T> message, CancellationToken cancellationToken);
-}
