--- src/Halcyon.Api/Common/Realtime/SignalRExtensions.cs
@@ -1,43 +1,0 @@
-﻿using System.Reflection;
-using System.Text.Json.Serialization;
-using Halcyon.Api.Common.Realtime;
-using Microsoft.AspNetCore.SignalR;
-
-namespace Halcyon.Api.Common.Realtime;
-
-public static class SignalRExtensions
-{
-    private static readonly MethodInfo _mapHubMethod = typeof(HubEndpointRouteBuilderExtensions)
-        .GetMethods(BindingFlags.Public | BindingFlags.Static)
-        .First(m => m.Name == "MapHub" && m.IsGenericMethodDefinition);
-
-    public static IHostApplicationBuilder AddSignalR(this IHostApplicationBuilder builder)
-    {
-        builder
-            .Services.AddSignalR()
-            .AddJsonProtocol(options =>
-            {
-                options.PayloadSerializerOptions.DefaultIgnoreCondition =
-                    JsonIgnoreCondition.WhenWritingNull;
-                options.PayloadSerializerOptions.Converters.Add(new JsonStringEnumConverter());
-            });
-
-        return builder;
-    }
-
-    public static WebApplication MapHubs(this WebApplication app, Assembly assembly)
-    {
-        var hubs = assembly
-            .GetTypes()
-            .Where(t => t.IsClass && !t.IsAbstract && typeof(Hub).IsAssignableFrom(t));
-
-        foreach (var hub in hubs)
-        {
-            var genericMethod = _mapHubMethod.MakeGenericMethod(hub);
-            var hubRoute = $"/hubs/{hub.Name}";
-            genericMethod.Invoke(null, [app, hubRoute]);
-        }
-
-        return app;
-    }
-}
