--- src/Halcyon.Api/Features/Account/ForgotPassword/ResetPasswordRequestedEvent.cs
@@ -1,3 +1,0 @@
-﻿namespace Halcyon.Api.Features.Account.ForgotPassword;
-
-public record ResetPasswordRequestedEvent(Guid UserId) { }
