--- src/Halcyon.Api/Features/Notifications/INotificationClient.cs
@@ -1,6 +1,0 @@
-﻿namespace Halcyon.Api.Features.Notifications;
-
-public interface INotificationClient
-{
-    Task ReceiveNotification(Notification notification, CancellationToken cancellationToken);
-}
