--- src/Halcyon.Api/Features/Notifications/Notification.cs
@@ -1,3 +1,0 @@
-﻿namespace Halcyon.Api.Features.Notifications;
-
-public record Notification(string Type, dynamic Data);
